+++ RELEASES.md
@@ -6,3 +6,4 @@
 
 - Added `Direction` (a normalized `Vec2`) and `Rotation` (an angle from midnight) to make it easier to work with rotations in 2D.
 - Added a `Position<T>` 2-dimensional coordinate type.
+- Added `TwoDimBundle` and `TwoDimPlugin` for working with and synchronizing 2D geometry within Bevy